refactor: vuex store -> provide/inject for report state
part of story #38263 Choose my own columns based on field name (numeric,
text, dates)

How to test:
- No functional change expected with or without the feature flag
  tuleap config-set feature_flag_enable_tql_select 1
- Editing the query, saving the report and cancelling both should
function as before.

Why?
We do not need a full Vuex (or Pinia) store for the Cross-tracker
widget, there are only a handful truly "global" pieces of data. Using
provide/inject can fully replace it.
The display of the CSV export button depends on Vuex store, my goal is
to fully replace it by provide/inject in order to avoid adding more code
that depends on Vuex store in the new SelectableTable.

Change-Id: Ie8e13d3f40b80c123de5ca133baa07e2ca76afed