refactor: vuex store -> provide/inject for feedbacks
part of story #38263 Choose my own columns based on field name (numeric,
text, dates)

How to test:
- All error cases should show a danger feedback with the error text:
  - fetching the report
  - fetching the projects of the user
  - fetching the trackers of the project
  - adding more than 25 trackers
  - fetching the artifacts matching the query
  - saving the report
  - exporting to CSV
- The CSV export button should still be displayed as before. It is only
shown when the report is saved, the user is not admin or the user is
admin and there are no invalid trackers, the table is not empty and not
loading.
- The success message is still shown after saving the report.
- No other functional change expected with or without the feature flag
  tuleap config-set feature_flag_enable_tql_select 1

Why?
We do not need a full Vuex (or Pinia) store for the Cross-tracker
widget, there are only a handful truly "global" pieces of data. Using
provide/inject can fully replace it.
The display of the CSV export button depends on Vuex store, my goal is
to fully replace it by provide/inject in order to avoid adding more code
that depends on Vuex store in the new SelectableTable.

Change-Id: I3dafea4f4d7b18ead685e75ae28375e7a320209e