refactor: vuex store -> provide/inject for report data
part of story #38263 Choose my own columns based on field name (numeric,
text, dates)

How to test:
- Set up a project A with at least one tracker (called T) in it.
- In another project B, set up a cross-tracker widget. Select the
tracker T.
- As site admin, mark the project A as suspended.
- Go back to the cross-tracker widget. As before, if you are widget
admin, it should show a danger alert explaining that some trackers are
invalid. Widget admins -> project admins on project dashboard, or -> the
owner of the user dashboard.
- Fetching the report, the report content, saving the report and CSV
export should function as before.
- As before, only widget admins can click on the widget to edit the
query and select trackers.
- No other functional change expected with or without the feature flag
  tuleap config-set feature_flag_enable_tql_select 1

Why?
We do not need a full Vuex (or Pinia) store for the Cross-tracker
widget, there are only a handful truly "global" pieces of data. Using
provide/inject can fully replace it.
The display of the CSV export button depends on Vuex store, my goal is
to fully replace it by provide/inject in order to avoid adding more code
that depends on Vuex store in the new SelectableTable.
It also reduces the weight of the app: 352.55 kB -> 337.81 kB.

Change-Id: I9cd2208144ab511dbb745e279e5e8d0b45b45980