feat: Add a request to save the query
Part of story #37941 Timetracking management widget

To avoid the query being updated in the database every time the 'Search' button is clicked, we have added a save request.
If the user switches to edit mode and clicks on 'Search', two buttons appear below the display mode.
A 'Save query' button to save the query in database for future use.
A 'Cancel' button for not saving the query.

The design of this feature is based on the Report save request in the Timetracking Overview widget.

The 'Save query' button is linked to the REST route.
The ISO format of the Date objects in javascript and the dates expected for the rest route are not the same.
2024-08-18T00:00:00.000Z
2024-08-18T00:00:00Z
This is why I use `replace(/.\d+Z$/g, "Z")` which removes the miliseconds part.

How to test:
-- Click on dates to switch to edit mode and click on the 'Search' button
-> The two new buttons should appear
-- Click on 'Save Query'
-> It should update your database with the corresponding values
-- Click on dates again to switch in edit mode and click on the 'Cancel' button
-> The two buttons should not appear

The widget is currently disabled by a feature flag. To set the feature flag to 1:
tuleap config-set feature_flag_allow_timetracking_management_widget 1

Change-Id: Ia8f67fef9a5f364e4ae39a9d8dcc2d900c375841