refactor: extendable-error -> Result
part of story #38263 Choose my own columns based on field name (numeric,
text, dates)

- When you try to select more than 25 trackers, it should show an error
message as before. You can edit the value in
writing-cross-tracker-report to ease the test.
- No functional change expected with or without the feature flag
  tuleap config-set feature_flag_enable_tql_select 1

Why?
We used extendable-error in 2017 when it was not safe to extend `Error`
and we did not know yet about Results. Using Result, we can better
signal that adding a tracker can fail, and using Fault, we can write a
dedicated translated message without direct dependency on the Fault
type.

Change-Id: If3172c755cc8fe256736942c449bcee5797ba746