feat: headers and footers
Note:

I couldn't find a way to hide the header and the footer on the first
page. These elements having a fixed position in order to be repeated on
each page, it is impossible to select only the ones on a given page.

Hiding them by making the first page content overlap them does not work
either because those elements being fixed, they are never affected by
z-indexes, and therefore, they are displayed on top of everything.

How to test:
- Click on the button to create a new template
--> There is a default header
--> There is a default footer

- Customise the header and the footer content
- Preview the result
--> The custom header and footer are displayed in the preview

- Save the template
- Duplicate the template
--> The new template inherits the custom header/footer defined
previously.

- save it and update it
--> The header and footer have been updated

Part of story #38630: Export to file

Change-Id: Iff02b4004bb0fa834a4e0d2aca457fbce1b04829