feat: Text values are not nullable
Part of story #38263 Choose my own columns based on field name (numeric,
text, dates)

How to test:
- Set up two trackers, one with a text field and the other without.
- In your cross-tracker search widget, write a query that SELECTs this
field with the two trackers
- No functional change expected: for artifacts of the tracker without
the text fields, cells should still be empty, there should be no error.

Why?
Defaulting to empty string is okay, since there is no need to
distinguish between "no value" or "intentional empty value". For dates
and numeric fields, we cannot assume a default value, so we must leave
the possibility to be null when there is no value or the field does not
exist in this tracker. Removing a null is a benefit because it
simplifies our code.

Change-Id: Ie6a60b2db4e312696cef1f9405a54e09787887e8