fix: Fix medium size modal and add wrappers for tabs and buttons
part of request #37211 Replace TLP-doc documentation

How to test:
- pnpm --filter=tlp-doc-v2 run storybook
- In TLP > Fly Over > Modals, choosing the control Size "medium-sized"
now works. It had a missing "d" and didn't change the modal.
- In TLP > Buttons & Switch > Buttons, the "classic" button is now
centered. The wide button is limited to 400px but it should be enough.
- In TLP > Structure & Navigation > Tabs, if you switch on the control
"With menu", the height of the canvas grows so that you can see the
dropdown menu without scrolling.
- In TLP > Fly Over > Popovers, the height of the canvas has been
increased so that when you switch on the control "With footer", the
popover still fits in the canvas without scrolling.
- No change for dropdowns and other modals.

Why?
All CSS classes are loaded together, so we must take care to not have
name collisions for storybook "wrapper" classes. They are all prefixed
now.
Position: relative and width 100% did not have an effect, so they were
removed.

Change-Id: Ib7ef007e82cf18daae70a2c1c5011652d7a7657a