fix: Artidoc TOC in PDF contains an extra "0."
In artidoc, the links inside the table of content have the same class
than the sections titles in the document.

In the pdf template admin, the default styles use this class (e.g
section-title) to prepend the number of the section to the section
title.

The proposition is to put a dedicated class on the table of content's
links, and let the list numbering do its job.

How to test:
- In artidoc reading mode
--> No visual change on the table of content nor sections titles

- In artidoc pdf export preview
--> No extra "0." is prepended to the toc items
--> The sections titles are still prepended with the section number

Part of story #38630: Export to file

Change-Id: Ib6250577b9a1142f32a83a8319b807ac7fc5dc2d