feat: ability to upload an image
Administrators can now upload an image in pdf template => images
library.

For now the images are not displayed so the best way to assert that the
upload really succeed is to look at the db table
`plugin_pdftemplate_image` and at the filesystem in
`/var/lib/tuleap/pdftemplate/images`.

Part of story #38630: Export to file

Change-Id: Ic317a9e75e970aee88745078ce34c825d181db45