feat: Ability to delete image
An image can now be deleted from PDF Templates Images Library.

If this image is detected as beeing used in some templates, then a
warning is displayed. To do that we detect if there is a reference of
the image href in the header, footer, or styling area of the templates.

Part of story #38630: Export to file

Change-Id: Id2c5add1d7ae909d013ed553b8e19f86ae4fd93e