feat: cover page in artidoc
The templates can now have a cover page.

Until now the cover page was static, user was not able to modify it,
therefore the title of the document was hardcoded. However the title
page can now have images or text above, below or around title. This
means that artidoc is not able anymore to introduce its document title.
We are therefore forced to introduce the concept of variables that
depends on the context of the print:
* artidoc exposes the document title
* template preview exposes an highlighted version of the variable

Note: since "cover page" can reference a front cover page (title) or a back
cover page (e.g. copyrights) we use internally the term [title page][0]
instead of cover page so that there will be less chance of conflict the
day we want to introduce the ability to have a back cover page.

Technical note: replacing variables, or injecting title page, header,
footer during impression should probably be delegated to print-as-pdf
internal library, so that we don't repeat ourselves each times we want
to export something as pdf. This will be done in a dedicated changeset.

[0] https://en.wikipedia.org/wiki/Title_page

Change-Id: I6425097e4101b2f9b3986c0205045789133c9347