feat: detect misusages of variables
If the user enters `${ DOCUMENT_TITLE }` instead of `${DOCUMENT_TITLE}`
then the variable won't be interpreted. In order to prevent such
situation, we now detect if the format is not correct.

Since the content is up to the user, in order to avoid false positive,
the save is not blocked if we detect misusages, only a warning is
raised.

Part of story #38630: Export to file

Change-Id: Ide1d0151703ea245f0c5cf81253fd3522872b6e2