fix: reset list-style-type for section content
Nested unnumbered list `ul`, by default have a disc at the first level,
a circle at the second level, then square for other levels.

Since we use a list to display the sections of a document, we should
reset the list-style-type so that we have the default inside the
document section.

We should have the following in every situation (prosemirror, ckeditor,
readonly, print):
```
●  first level
   ○  second level
      ■  third level
```

Discovered during story #38627: Choice of a new Rich Text Editor

Change-Id: I1d1f6eb9642bb330b27aed06ef4de923b362211c