fix: escape variable interpretation
We should not allow variable interpretation inside PDF template
generation.

Given an artidoc with title "Document <br> title"
When we export it as PDF
Then the title should display:
"Document <br> title"
And not:
"Document
title"

This forces us to ditch the variable visualisation in PDF preview.

Part of story #38630: Export to file

Change-Id: I6ae00c4b5c57848924efd729c8f00423c0c0eb4b