fix: User list fields can have user 100 in database
Part of story #38310 Choose my own columns based on field name (lists)

For multi select box and checkbox no values is stored as value 100 in
database, or 100 is user None so when we get user list we get this user
but we should not.

To avoid array -> object conversion when json_encoding API result, add
use of array_values on each list representation

*Testing:*

Have a tracker with a multi select box bind to users and in an artifact
left this field empty. If you select on this field in your XTS widget
the API return should be empty for this field

Change-Id: I3c07d5617c2c789e3507ad3d45e1ea43c658bcd2