chore: DB test for XTS report DAO
part of request #40099 Cross-Tracker search finishing touches 16.1

No functional change. DB integration tests should pass.

Why?
In order to properly duplicate the widget data when inheriting from
another project, we will need to touch this DAO. The tests will ensure
non-regression.
Note that I still had to change to tryFlatTransaction() in the same
change, because the tests start a transaction and then roll it back.
Without this change, the tests fail because a transaction is already
ongoing.

Change-Id: Ib061109834e9134e8c7072feff46ac98bc238ddf