feat: Closing editor should stop ongoing file uploads
When I drop a bunch of files in the section editor, and then cancel the
edition of this section, then the upload of the file should stop
(corresponding patch is marked as canceled in the dev tools) and the
progress bar should disappear.

The upload is aborted but not terminated, so that if the user drop again
the file, then the already uploaded bits are kept and new upload will
resume faster.

Part of story #38627: Choice of a new Rich Text Editor

Change-Id: I0d083632f4614b28e6b9749eaea861e288ea7b27