feat: Harmonize link color and center xref badges
Part of story #39019 Switch mode in Cross tracker search widget

How to test:
- Using the API explorer, modify (PUT) a cross-tracker report to be in
"expert" mode.
- Write a query so that it selects @pretty_title and lists bound to
users.
- The "Edit" link in the first column, the link in the pretty title and
the links to user profiles in the list columns should all be black. When
you hover them or focus them (using the keyboard), they become
tlp-main-color (depending on your theme).
- The cross-ref badge in pretty title should be more centered in
relation to the text, when the text does not wrap.
- Go to a widget in "default" mode, select some artifacts.
- The cross-ref badge in the "pretty title" should be more centered in
relation to the text. The link should be black and tlp-main when hovered
or focused.
- No visual change expected on the Assigned to, Submitted by, Last
update date columns. The links should have the same colors.

Note:
- Heartbeat widget links might change, because they were (wrongly)
targeted by the CSS class in cross-tracker-widget, so on dashboard with
both widgets, the links showed differently. The CSS is now scoped by
Vue, so it should not cross-target anymore.

Change-Id: I0fa52dd34e42765c1169d2be0e230e45b8548169