Fix: Whitespace adjustments
Instead of recreating manually sections in a pane with padding and
border, we should use existing tlp-pane-section. Since cross tracker
widget needs to have more than one section (one for the request and one
for the results) we have to disable the default tlp-pane-section that is
added automatically for each widget by the dashboard.

Form elements have a tlp-medium-spacing between them, so trackers and
query should have this spacing between them. Furthermore, the trackers
list was closer to the request than to the trackers selection.
Whitespace between selection and list is reduced so that they are
logicaly grouped together.

On a table row, the first and last cells should have the same padding
left and right than the rest of the tlp-pane-section
(--tlp-medium-spacing).

On reading mode, when we hover the report, there should be some
whitespace between the background color and the text.

Success and error messages don't have anymore too much whitespace around
them.

Not a padding adjustment, but success message is now marked as success
instead of info.

Part of story #39019: Switch mode in Cross tracker search widget

Change-Id: Ifd44f448e0a7e5ed9f8c156408e7f15d0a43cdec