fix: force full page width
When an artidoc does not have much content, the content appears to be
shrinked. This is due to various css hacks used for header and footer.

The proposal is to include a wide image to force the content to occupy
the full page width.

Part of story #38630: Export to file

Change-Id: Ia92c54cff02ec02041490637a1e57206d632269b