refactor: Add interface before list field retrieval
No functional change. Unit tests should pass.

Why?
This series of refactorings has the goal to make it easy to write a
"domain" API to find the current column of a given card. This requires
us to know which field in the card's tracker is "mapped" (either
freestyle or status semantic). We start by simplifying the freestyle
mapping part.

Change-Id: I3c58e90ec65d1fa9ee8be4bee58f8383a359d0c7