Fix request #39332: filename-pattern flaky test
Sometimes filename-pattern e2e test is failing because it counts the
number of rows in a table, asserting that there are 6 rows (5 for each
uploaded file during the test + 1 header row). However sometimes it
count the skeleton instead and since there are 5 skeletons it thinks
that it is ok.

Moving the data-test attribute from the table to the real row will
prevent this issue to appear again.

No functional changes, e2e tests should pass.

Change-Id: I3bf1ca95eb9149eb89e1a78df2ffa2338788542b