refactor: Split mapped field retrieval out of FreestyleMappingFactory
part of request #39022 Moving a card in Taskboard can have side effects

No functional change. Unit, REST and DB integration tests should pass.

Why?
This series of refactorings has the goal to make it easy to write a
"domain" API to find the current column of a given card. This requires
us to know which field in the card's tracker is "mapped" (either
freestyle or status semantic). We start by simplifying the freestyle
mapping part.

FreestyleMappingFactory did not have clear responsibilities. It is now
split in two: FreestyleMappedFieldRetriever (that focuses on the mapped
field), and the rest for values. The factory itself will be renamed
later.

Added interfaces before DB queries. Interfaces will allow us to make the
DAO final and write stubs to replace it in unit tests.

Also removed static build methods. The final, top-most endpoints should
have the responsibility of instantiating everything, so that they can
substitute interfaces with caches if needed. Instantiation should not be
a responsibility of individual objects.

Change-Id: I61881f00bf082fe074ac842d997a0779de59b8e5