refactor: rename FreestyleMappingFactory
part of request #39022 Moving a card in Taskboard can have side effects

No functional change. Unit, REST and DB integration tests should pass.

Why?
This series of refactorings has the goal to make it easy to write a
"domain" API to find the current column of a given card. This requires
us to know which field in the card's tracker is "mapped" (either
freestyle or status semantic). We start by simplifying the freestyle
mapping part.

FreestyleMappingFactory did not have clear responsibilities. It is now
renamed into FreestyleMappedFieldValuesRetriever (that focuses on the
values).

Added interfaces before DB queries. Interfaces allow us to make the DAO
final and write stubs to replace it in unit tests.

Also removed static build methods. The final, top-most endpoints should
have the responsibility of instantiating everything, so that they can
substitute interfaces with caches if needed. Instantiation should not be
a responsibility of individual objects.

Change-Id: Id7c9c03f4e9e82ae3be8e2bff3e4ca29baa8b2f2