Fix homepage display in big screens
When the width of the viewport is bigger than 3000px, then the page
appears broken.
This is due to the illustration svg image that is hard coded to 3000px.

As a quick and dirty fix we now scale the illustration for viewport
bigger than 3000px.

Furthermore, when welcome text is floating over the background
illustration, then it is not readable at all. By applying a text-shadow
we manage to have something more readable than before.

This is a quick and dirty fix. A better approach would be to rewrite how
the background image is included and be able to use `mix-blend-mode:
difference` to have a readable text without text-shadow. This might need
to change the background image or the design of the page.

Fixes request #39338 Homepage display is broken on large screens

Change-Id: I7c0b7c7491fafd6c2bb56683b33619958aeff786