refactor: request #39344 DatabaseUUIDFactory::buildUUIDFromHexadecimalString returns an Option<UUID>
It currently throws when the UUID cannot be decoded.
This design will likely lead to the possibility for end-users to
generate crashes by providing an invalid string. This is not desirable.

An Option forces the developer to make an explicit decision about it.
As the decision is likely to be "do nothing" like when the provided ID
does not exist in the DB, an Option seems more suitable than Ok|Err.

No fonctional changes expected.

Change-Id: I67e399a912f01bdfa524cc7b7efa16520f01f8bd