refactor: MappedFieldRetriever returns Option
part of request #39022 Moving a card in Taskboard can have side effects

No functional change. Unit and REST tests should pass.

Why?
This series of refactorings has the goal to make it easy to write a
"domain" API to find the current column of a given card. This requires
us to know which field in the card's tracker is "mapped" (either
freestyle or status semantic). We continue by simplifying the retrieval
of the mapped field.

Change-Id: I01e618e033f0ad3f96f4108466f91ba855bd8ae0