feat: Check @project = 'self'
Part of story #39330 Choose on which tracker I perform my query

Add an interface to ProjectDashboardDao which tries to retrieve the
project id from the widget

Then use it in validator to check if currrent XTS widget is in a
project. If not so add an error to the error collector as @project =
'self' has no meaning outside a project.

@project = 'aggregated' is marked as not supported yet

The implied @project = 'self' when there is only @tracker condition will
be managed from the @tracker point of view

*Testing:*

- In your project dashboard XTS widget, you can use @project = 'self'
- In your personnal dashboard XTS widget, there is an error message
  shown

Change-Id: I5733b6860a4cc30e898876147103b2d177988461