refactor: MappedValuesRetriever returns Option
part of request #39022 Moving a card in Taskboard can have side effects

No functional change. Unit and REST tests should pass.

Why?
This series of refactorings has the goal to make it easy to write a
"domain" API to find the current column of a given card. This requires
us to know which list bind values in the card's "mapped" field are
associated to a given column. We continue by simplifying the retrieval
of the mapped values.

Change-Id: I212bc217b8591a3c5733ee3e1aef05f6bb36e361