refactor: ArtifactMappedFieldValueRetriever returns Option
part of request #39022 Moving a card in Taskboard can have side effects

No functional change. Unit and REST tests should pass.

Why?
This series of refactorings has the goal to make it easy to write a
"domain" API to find the current column of a given card. This requires
us to know which list bind values have been selected in the last
changeset of a card's artifact. We continue by simplifying the retrieval
of the bind values of a card.

Also removed static build methods. The final, top-most endpoints should
have the responsibility of instantiating everything, so that they can
substitute interfaces with caches if needed. Instantiation should not be
a responsibility of individual objects.

Change-Id: Iafec03909f7d1023763d30d540b04ecfe6a44beb