fix: ignore update of card already in its column
Closes request #39022 Moving a card in Taskboard can have side effects

How to test:
- Create a new Scrum project. The default "Bug" tracker will be used.
- Go to the Bug tracker administration > Workflow > Transition rules.
- Enable the transition rules.
- Create a new Release artifact.
- Add a new Bug artifact to the Release backlog, choose "To do" status.
- Open a first browser tab to the Release taskboard.
- Open a second browser tab to the same taskboard.
- In the first tab, move the card to the "Done" column.
- Open in a third browser tab the artifact view of the Bug card. => The
Status value should be "Invalid".
- In the second tab, without refreshing, move again the card to the
"Done" column.
=> In the artifact view, after refreshing, the status value should still
be "Invalid".
=> Before the fix, the status value would be "Wont fix".

Why?
When moving a card to a column where it already is, the update process
of the card artifact is triggered. This can happen with concurrent
updates (two people moving the same card to the same column). It may
create an unwanted changeset, and may trigger unwanted workflow
transition or post actions. Also, when multiple values map to a single
column, it will cycle through the values and select the next value.

Instead of doing the update, we should do nothing: the card is already
in the given column, so there is nothing more to do.

Also removed static build methods. The final, top-most endpoints should
have the responsibility of instantiating everything, so that they can
substitute interfaces with caches if needed. Instantiation should not be
a responsibility of individual objects.

Change-Id: I0e515dbdbdb57bd4b953c6dcef13d3c82ec18e8e