chore: namespace ColumnCollection and ColumnFactory
part of request #39022 Moving a card in Taskboard can have side effects

No functional change. Unit and REST tests should pass.

Why?
This series of refactorings has the goal to make it easy to write a
"domain" API to find the current column of a given card. This requires
us to know the list of columns of the affected taskboard.

Removed the two inheritor classes of ColumnCollection, as the visitor
they implemented was not called anymore.

Change-Id: If3c24c8bb6a71a85ce0960b74f9b86198c5fc275