feat: Show PR branches on homepage
Implements request #38641 pull request dashboard should display branches

How to test:
- prepare several pull requests: one that comes from a fork, one that is
in the same repository
- prepare a broken pull request. You can forge the status (broken = 2)
in the plugin_pullrequest_git_reference DB table.
- Browse the list of pull requests in the destination repository. You
should see the source and destination branches of all pull requests. The
one coming from the fork has the fork name as prefix. The broken pull
request's branch badge is still aligned with the others.
- Browse the list of pull requests in the fork. The pull request going
to the origin repo has the repo name as prefix.

Unless extremely long branch names are used, the branch badges should be
aligned on their right side. The labels and reviewers parts have a max
width and can shrink. The pull request title can shrink or grow until it
reaches the branch badge.

The first line of the title, the branch badge, the reviewers list and
the first label badge are centered vertically.

The branch badge in the overview tab should also have space around the
arrow icon.

Change-Id: I28a617fca440ed2ba277021623775daad5449fc6