feat: add upload global store and multiple progress bars
When you're uploading a bunch of files, a list of progress bars is displayed on top of the page as temporary alerts.
To do that, I had to refactor the state managment for upload. The state was stored by section, and now it has to be in a global store.
This new approach fixed the bug when the save button was enabled between uploads.

part of story #38627 : Choice of a new Rich Text Editor

Change-Id: I9bb697669d9ae085e277e0465ce4f3c926dd7566