feat: Build @tracker.name SQL query
Part of story #39330 Choose on which tracker I perform my query

You can now search trackers by their name with @tracker.name

= 'some' and IN() will match all trackers with the corresponding short
name. The match is exact

If @project condition is not provided, it will default to @project =
'self'. So in the context of a personal dashboard it leads to an error
(project condition is mandatory).

*Testing:*

Have some project with same category and in each a tracker with the same
short name.

With the query: `SELECT @project.name, @tracker.name FROM
@project.category = 'you_cat' AND @tracker.name = 'your_tracker' WHERE
@id >= 1` you should get all corresponding trackers.

The query should also works with `@tracker.name IN(...)`

Change-Id: I3572f7ca3fa755f7c79921ce22f778c37376d390