refactor: Extract cache out of ProgramAdapter
part of request #39352 Program add to top backlog post-action should
ignore Teams

No functional change. The number of DB queries to check that a user can
access a Program should stay the same.

Why?
Caches should leverage interfaces. By making a cache implement the same
interface and "wrap" the real implementation, we can substitute it
everywhere it's needed. It also splits responsibilities: managing a
cache is not the same as "doing the actual stuff". It's also much easier
to unit test when no cache is involved.

This also inlines a bypass parameter that was always null.

Change-Id: Ie0c06a8dd063a1b4718b8dd8860f9e94a701d886