feat: Use different program check to list top backlog actions
part of request #39352 Program add to top backlog post-action should
ignore Teams

How to test:
- Create a new Program project from the Agile Release Train template.
- Do not link any Teams to the Program.
- Go to the Features tracker administration > Workflow > Transitions.
When you click on the "To Plan" state config, you should see the "Add to
top backlog" post-action, but its name is blank. Same for "Scheduled"
and "WIP" columns.
- If you try to trigger the post-action, it will continue doing nothing.
This will be addressed in a separate commit.
- Without this patch, in the same columns, the post-action was not
listed at all (the REST route returned []).

Why?
The existence of the "Add to Top backlog" post-action was linked to
having a Program completely configured and linked to Teams. There is no
real functional reason to forbid this post action when the Program has
no Teams. At worst case, it will add to the top backlog, but then it
cannot be planned anywhere because there are no Teams.
This makes it hard to use Templates, because by nature, a Template of a
Program will never be linked to any Teams, as its purpose is to be
duplicated. When this duplication occurs, the code checks again to see
if there are any Teams linked to the template, and as a result, the
workflow cannot be inherited.
This is a first step towards fixing this duplication.

Change-Id: Ieee032616a8db8bb4c2851ba154da76f2a18aa18