refacto: Use CrossTrackerReportDao::searchCrossTrackerWidgetByCrossTrackerReportId
Part of story #39330 Choose on which tracker I perform my query

In a previous I've added a method to ProjectDasboardDao to retrieve
project id from the widget_content_id and widget_type. But there was
already an existing one in CrossTrackerReportDao.

This commit fix this duplication by using the existing method and
removing 🔥 the one from ProjectDashboardDao (not used anymore)

*No functional changes expected*

Change-Id: If72b0e4b688022c131355c43c963af1e82e0c58c