feat: create section from existing artifacts
User can now insert a section from an existing artifact. A modal is
displayed to allow to search for artifact, based on its title.

Primary behavior for empty document is still to create new sections from
scratch. The pending artifact section is still displayed as before. In a
future contribution this pending artifact section will be removed if
user decides to insert an existing one (if this pending section is still
empty).

For now if current user cannot create a new section, then current user
cannot add an existing one as well. This will change in a future
contribution as we want to allow them to be able to insert an existing
section even if they cannot create new sections. They must have write
permission on the document though.

It is not possible to select more than one artifact to insert many
sections in one shot.

part of ???

Change-Id: I52763fd8d91bd7e042a84df49bc980543ac5af97