feat: add notification message and upgrade UX
Create a notification store to display info/error notification on top of the page
Fix style overflow text when file name is too long.
Limit the number of displayed notifications simultaneously, add a message to indicate the number of remaining upload in progress.

part of story #38627 : Choice of a new Rich Text Editor

Change-Id: I21bc46b4066b9aa4058ca6287d2bddaf7c29d227