feat: ability to copy link url
part of story #38627 : Choice of a new Rich Text Editor

Note: I had to downgrade hybrids to 9.1.1 because of a breaking change
introduced in version 9.1.2 making the component crash when it passes
properties to children components. I'll try to find a way around during
the next big hybrids upgrade in the whole codebase (other hybrids
components use 9.0.0).

How to test:
- Click on a link in an editor
--> A popover menu opens
--> It contains a "copy to clipboad" button

- Click it
--> The url has been copied to clipboard
--> The icon of the button changes to fa-check for 2s before changing
back to fa-copy.

Change-Id: Ia733bdebc6c383402a210334c5d947674552c66e