feat: Add highlight for SELECT FROM WHERE
Part of story #39330 Choose on which tracker I perform my query

These keywords should be highlighted because they are important.

FROM was already highlighted because of 'IS LINKED FROM' and its
friends. Add it by its own to avoid loosing it

*Testing:*

When typing SELECT/FROM/WHERE autocompletion is possible and the word is
highlighted

Change-Id: I2fa476682a9af166821d22227382e4ce5485a352