feat: ATF field is not mandatory to use it in XTS
Closes request #39368 Always there field existence is not mandatory to use it in XTS TQL

Until now, to use an always there field in the condition of your query,
your trackers must have this field defined. In a way it breaks the fact
that the field is 'always there'. So now if the field doesn't exist, we
keep the tracker. But if the field exists we check the permissions on
it.

*Testing:*

- If your tracker doesn't have the field you should be able to query on
  it
- If your tracker has the field, it depends on the permissions of the
  field

Change-Id: I94a27715850d2a7f28b33cc743a3ea391e491145